Move quick reply models into separate file
